pyproject.toml
[project]
name = "mindpulse-endpoint-poc"
version = "0.1.0"
description = "Proof-of-concept endpoint for the MindPulse Android screenshot app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-flask>=1.3.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
mindpulse-endpoint = "mindpulse_endpoint_poc.app:create_app"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
